@@ -22,10 +22,12 @@
     "test:watch": "jest --watch",
     "postinstall": "opencollective-postinstall || exit 0",
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
-    "prettify": "prettier --single-quote --trailing-comma=es5 --write './**/*.md'",
     "clean-install": "rimraf node_modules && yarn",
     "changelog": "auto-changelog -p",
-    "docs": "cd website && yarn start"
+    "docs": "yarn docs:build && cd website && yarn start",
+    "docs:lint": "prettier --write '**/*.md'",
+    "docs:build": "ts-node website/scripts/docgen/index.ts && yarn docs:lint",
+    "docs:test": "cd website && jest --runInBand"
   },
   "author": "Nader Dabit & Monte Thakkar",
   "license": "MIT",
@@ -52,7 +54,6 @@
     "@react-native-community/eslint-config": "^2.0.0",
     "@testing-library/react-native": "^7.0.2",
     "@types/color": "^3.0.1",
-    "@types/enzyme": "^3.10.8",
     "@types/hoist-non-react-statics": "^3.3.1",
     "@types/jest": "^26.0.23",
     "@types/lodash.isequal": "^4.5.5",
@@ -60,10 +61,8 @@
     "@types/react-test-renderer": "^17.0.0",
     "auto-changelog": "^2.2.1",
     "babel-jest": "^26.3.0",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.4",
-    "enzyme-to-json": "^3.5.0",
     "eslint": "^7.9.0",
+    "child_process": "^1.0.2",
     "husky": "^4.3.0",
     "jest": "^26.4.2",
     "jest-transform-stub": "^2.0.0",
@@ -76,7 +75,8 @@
     "react-native-vector-icons": "^7.0.0",
     "react-test-renderer": "^16.13.1",
     "rimraf": "^3.0.2",
-    "typescript": "^4.1.3",
+    "ts-node": "^10.2.0",
+    "typescript": "^4.3.5",
     "utility-types": "^3.10.0"
   },
   "peerDependencies": {
@@ -90,6 +90,7 @@
     "testPathIgnorePatterns": [
       "./src/SearchBar/__tests__/common.tsx",
       "<rootDir>/node_modules",
+      "<rootDir>/website",
       "<rootDir>/dist"
     ],
     "coveragePathIgnorePatterns": [
@@ -143,7 +144,8 @@
   },
   "husky": {
     "hooks": {
-      "pre-commit": "lint-staged"
+      "pre-commit": "lint-staged",
+      "pre-push": "node updateDocumentation.js"
     }
   }
 }
