@@ -1,16 +1,39 @@
-import React from 'react';
-import { render } from '@testing-library/react-native';
-import { ThemeProvider, FullTheme } from '../src/config';
+import React, { JSXElementConstructor } from 'react';
+import { View } from 'react-native';
+import {
+  render,
+  RenderOptions,
+  fireEvent,
+  act,
+} from '@testing-library/react-native';
+import { ThemeProvider, FullTheme, colors } from '../src/config';
+import deepmerge from 'deepmerge';
 
-const theme = {
-  colors: {
-    black: 'black',
-  },
-};
+export { fireEvent, act };
 
-export const renderWithTheme = (
-  children: any,
-  themeProp: Partial<FullTheme> = theme
+// for getting findByType e.g. wrapper.findByType(Icon) see implementation in Avatar Component
+export const renderWithWrapper = (
+  children: React.ReactElement<any, string | JSXElementConstructor<any>>,
+  wrapperTestID?: string,
+  themeProp: Partial<FullTheme> = {},
+  renderOptions?: RenderOptions
 ) => {
-  return render(<ThemeProvider theme={themeProp}>{children}</ThemeProvider>);
+  const options: RenderOptions = {
+    ...(!wrapperTestID && {
+      wrapper: (props) => <View {...props} testID="wrapper" />,
+    }),
+    ...renderOptions,
+  };
+  const renderApi = render(
+    themeProp ? (
+      <ThemeProvider theme={deepmerge({ colors }, themeProp)}>
+        {children}
+      </ThemeProvider>
+    ) : (
+      children
+    ),
+    options
+  );
+  const wrapper = renderApi.queryByTestId(wrapperTestID || 'wrapper');
+  return { wrapper, ...renderApi };
 };
