@@ -7,22 +7,37 @@ import {
   StyleProp,
   ViewStyle,
   ModalProps,
+  ScrollViewProps,
 } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { RneFunctionComponent } from '../helpers';
 
 export type BottomSheetProps = {
+  /** Style of the bottom sheet's container. Use this to change the color of the underlay. */
   containerStyle?: StyleProp<ViewStyle>;
+
+  /** Additional props handed to the `Modal`. */
   modalProps?: ModalProps;
+
+  /** Is the modal component shown. */
   isVisible?: boolean;
+
+  /** Used to add props to Scroll view. */
+  scrollViewProps?: ScrollViewProps;
 };
 
+/** Overlay Modal that displays content from the bottom of the screen.
+ * This opens from the bottom of the screen.
+ * **Note:**
+ * Make sure that you have completed [Step 3](getting_started.md#step-3-setup-react-native-safe-area-context) in the setup guide before using `BottomSheet`.
+ */
 export const BottomSheet: RneFunctionComponent<BottomSheetProps> = ({
   containerStyle,
   isVisible = false,
   modalProps = {},
   children,
-  ...props
+  scrollViewProps = {},
+  ...rest
 }) => {
   return (
     <Modal
@@ -36,10 +51,10 @@ export const BottomSheet: RneFunctionComponent<BottomSheetProps> = ({
           styles.safeAreaView,
           containerStyle && containerStyle,
         ])}
-        {...props}
+        {...rest}
       >
         <View>
-          <ScrollView>{children}</ScrollView>
+          <ScrollView {...scrollViewProps}>{children}</ScrollView>
         </View>
       </SafeAreaView>
     </Modal>
