@@ -1,16 +1,31 @@
 import React from 'react';
-import { StyleSheet, Animated } from 'react-native';
+import { StyleSheet, Animated, StyleProp, ViewStyle } from 'react-native';
 import Button, { ButtonProps } from '../Button';
 import { RneFunctionComponent } from '../helpers';
 
 export type FABProps = ButtonProps & {
+  /** Change the color of the FAB. */
   color?: string;
+
+  /** Change Size of FAB. */
   size?: 'large' | 'small';
+
+  /** FAB placement at bottom, (optional) use [`style`](#style) in case of custom placement. */
   placement?: 'left' | 'right';
+
+  /** Decide the visibility of the FAB. */
   visible?: boolean;
+
+  /** Transform Extended Label text to uppercase. */
   upperCase?: boolean;
+
+  /** Style for FAB */
+  style?: StyleProp<ViewStyle>;
 };
 
+/** A floating action button (FAB) performs the primary, or most common, action on a screen.
+ * It appears in front of all screen content, typically as a circular shape with an icon in its center.
+ * Also receives all [Button](https://reactnativeelements.com/docs/button#props) props. */
 export const FAB: RneFunctionComponent<FABProps> = ({
   color,
   size = 'large',
@@ -24,7 +39,7 @@ export const FAB: RneFunctionComponent<FABProps> = ({
   containerStyle,
   iconContainerStyle,
   placement,
-  ...props
+  ...rest
 }) => {
   const { current: animation } = React.useRef(
     new Animated.Value(Number(visible))
@@ -40,7 +55,7 @@ export const FAB: RneFunctionComponent<FABProps> = ({
 
   return (
     <Animated.View
-      style={[
+      style={StyleSheet.flatten([
         {
           opacity: animation,
           transform: [{ scale: animation }],
@@ -53,11 +68,11 @@ export const FAB: RneFunctionComponent<FABProps> = ({
           bottom: 0,
         },
         style,
-      ]}
+      ])}
     >
       <Button
         buttonStyle={StyleSheet.flatten([
-          props.title
+          rest.title
             ? styles.extendedLabel
             : size === 'small'
             ? styles.smallFAB
@@ -68,7 +83,7 @@ export const FAB: RneFunctionComponent<FABProps> = ({
           buttonStyle,
         ])}
         iconContainerStyle={[
-          props.title
+          rest.title
             ? {}
             : size === 'small'
             ? styles.smallFAB
@@ -80,13 +95,14 @@ export const FAB: RneFunctionComponent<FABProps> = ({
           disabled && styles.disabled,
           containerStyle,
         ])}
+        /** For Extended FAB */
         titleStyle={[
           styles.label,
           { color: theme?.colors?.white },
           upperCase && styles.upperCaseLabel,
           titleStyle,
         ]}
-        {...props}
+        {...rest}
         {...{ disabled, theme }}
       />
     </Animated.View>
