@@ -1,14 +1,12 @@
 import React from 'react';
 import {
   Platform,
-  TouchableHighlight,
   View,
   StyleSheet,
-  TouchableNativeFeedback,
   ViewStyle,
   StyleProp,
   TextStyle,
-  TouchableHighlightProps,
+  Pressable,
 } from 'react-native';
 import {
   IconButtonProps,
@@ -17,7 +15,11 @@ import {
 import Color from 'color';
 import getIconType from '../helpers/getIconType';
 import getIconStyle from '../helpers/getIconStyle';
-import { RneFunctionComponent } from '../helpers';
+import {
+  androidRipple,
+  InlinePressableProps,
+  RneFunctionComponent,
+} from '../helpers';
 
 export type IconType =
   | 'material'
@@ -34,30 +36,62 @@ export type IconType =
   | 'font-awesome-5'
   | string;
 
-export type IconObject = TouchableHighlightProps & {
+export type IconObject = {
+  /** Name of icon. */
   name?: string;
+
+  /** Color of icon. */
   color?: string;
+
+  /** Size of icon. */
   size?: number;
+
+  /** Type of icon */
   type?: IconType;
+
+  /** Apply style to the icon using iconStyle. */
   iconStyle?: StyleProp<TextStyle>;
 };
 
 export type IconNode = boolean | React.ReactElement<{}> | Partial<IconProps>;
 
 export type IconProps = IconButtonProps & {
+  /** Type of icon set. [Supported sets here](#available-icon-sets). */
   type?: IconType;
-  Component?: typeof React.Component;
-  reverse?: boolean;
-  raised?: boolean;
-  containerStyle?: StyleProp<ViewStyle>;
-  iconProps?: VectorIconProps;
-  reverseColor?: string;
-  disabled?: boolean;
-  disabledStyle?: StyleProp<ViewStyle>;
-  solid?: boolean;
-  brand?: boolean;
-};
 
+  /** Update React Native Component. */
+  Component?: typeof React.Component;
+
+  /** Reverses color scheme. */
+  reverse?: boolean;
+
+  /** Adds box shadow to button. */
+  raised?: boolean;
+
+  /** Add styling to container holding icon. */
+  containerStyle?: StyleProp<ViewStyle>;
+
+  /** Provide all props from react-native Icon component. */
+  iconProps?: VectorIconProps;
+
+  /** Specify reverse icon color. */
+  reverseColor?: string;
+
+  /** Disables onPress events. Only works when `onPress` has a handler. */
+  disabled?: boolean;
+
+  /** Style for the button when disabled. Only works when `onPress` has a handler. */
+  disabledStyle?: StyleProp<ViewStyle>;
+
+  /** Uses the solid font. */
+  solid?: boolean;
+
+  /** Uses the brands font (FontAwesome5 only). */
+  brand?: boolean;
+} & InlinePressableProps;
+
+/** Icons are visual indicators usually used to describe action or intent.
+ * They are also used for displaying information. */
 export const Icon: RneFunctionComponent<IconProps> = ({
   type = 'material',
   name,
@@ -73,43 +107,38 @@ export const Icon: RneFunctionComponent<IconProps> = ({
   disabled = false,
   disabledStyle,
   onPress,
-  Component = onPress
-    ? Platform.select<typeof React.Component>({
-        android: TouchableNativeFeedback,
-        default: TouchableHighlight,
-      })
+  onLongPress,
+  onPressIn,
+  onPressOut,
+  Component = onPress || onLongPress || onPressIn || onPressOut
+    ? Pressable
     : View,
   solid = false,
   brand = false,
   theme,
-  ...attributes
+  pressableProps,
+  ...rest
 }) => {
   const color = colorProp || theme?.colors?.black;
   const reverseColor = reverseColorProp || theme?.colors?.white;
   const IconComponent = getIconType(type);
   const iconSpecificStyle = getIconStyle(type, { solid, brand });
 
-  const getBackgroundColor = () => {
+  const getBackgroundColor = React.useMemo(() => {
     if (reverse) {
       return color;
     }
     return raised ? theme?.colors?.white : 'transparent';
-  };
+  }, [color, raised, reverse, theme?.colors?.white]);
 
-  const buttonStyles = {
-    borderRadius: size + 4,
-    height: size * 2 + 4,
-    width: size * 2 + 4,
-  };
-
-  if (Platform.OS === 'android' && !attributes.background) {
-    if (Platform.Version >= 21) {
-      attributes.background = TouchableNativeFeedback.Ripple(
-        Color(color).alpha(0.2).rgb().string(),
-        true
-      );
-    }
-  }
+  const buttonStyles = React.useMemo(
+    () => ({
+      borderRadius: size + 4,
+      height: size * 2 + 4,
+      width: size * 2 + 4,
+    }),
+    [size]
+  );
 
   return (
     <View
@@ -125,30 +154,42 @@ export const Icon: RneFunctionComponent<IconProps> = ({
           : {},
         containerStyle && containerStyle,
       ])}
+      testID="RNE__ICON__CONTAINER"
     >
       <Component
-        {...attributes}
-        {...(onPress && {
+        {...{
+          android_ripple: androidRipple(
+            Color(reverse ? color : (underlayColor as string))
+              .alpha(0.3)
+              .rgb()
+              .string()
+          ),
           onPress,
+          onLongPress,
+          onPressIn,
+          onPressOut,
           disabled,
-          underlayColor: reverse ? color : underlayColor,
-          activeOpacity: 0.3,
-        })}
+          accessibilityRole: 'button',
+          ...pressableProps,
+          ...rest,
+        }}
+        testID="RNE__ICON__CONTAINER_ACTION"
       >
         <View
           style={StyleSheet.flatten([
             (reverse || raised) && buttonStyles,
             {
-              backgroundColor: getBackgroundColor(),
+              backgroundColor: getBackgroundColor,
               alignItems: 'center',
               justifyContent: 'center',
             },
             disabled && styles.disabled,
             disabled && disabledStyle,
           ])}
+          testID="RNE__ICON"
         >
           <IconComponent
-            testID="iconIcon"
+            testID="RNE__ICON__Component"
             style={StyleSheet.flatten([
               { backgroundColor: 'transparent' },
               iconStyle && iconStyle,
