@@ -33,7 +33,7 @@ export type InputProps = React.ComponentPropsWithRef<typeof TextInput> & {
   rightIcon?: IconNode;
   rightIconContainerStyle?: StyleProp<ViewStyle>;
   inputStyle?: StyleProp<TextStyle>;
-  InputComponent?: typeof React.Component;
+  InputComponent?: React.ComponentType | React.ForwardRefExoticComponent<any>;
   errorProps?: object;
   errorStyle?: StyleProp<TextStyle>;
   errorMessage?: string;
@@ -115,7 +115,10 @@ export class Input extends React.Component<
     const hideErrorMessage = !renderErrorMessage && !errorMessage;
 
     return (
-      <View style={StyleSheet.flatten([styles.container, containerStyle])}>
+      <View
+        testID="RNE__Input__view-wrapper"
+        style={StyleSheet.flatten([styles.container, containerStyle])}
+      >
         {renderText(
           label,
           { style: labelStyle, ...labelProps },
