@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import {
-  TouchableOpacity,
+  Pressable,
   LayoutAnimation,
   StyleSheet,
   View,
   ActivityIndicator,
   Text,
-  TouchableOpacityProps,
+  PressableProps,
   ViewStyle,
   StyleProp,
   TextStyle,
@@ -36,7 +36,7 @@ const defaultClearIcon = (theme: Theme) => ({
 export type SearchBarIosProps = InputProps &
   SearchBarBaseProps &
   typeof SearchBarIOS.defaultProps & {
-    cancelButtonProps?: Partial<TouchableOpacityProps> & {
+    cancelButtonProps?: Partial<PressableProps> & {
       buttonStyle?: StyleProp<ViewStyle>;
       buttonTextStyle?: StyleProp<TextStyle>;
       color?: string;
@@ -166,6 +166,7 @@ export class SearchBarIOS extends Component<
 
     return (
       <View
+        testID="RNE__SearchBar-wrapper"
         style={StyleSheet.flatten([
           styles.container,
           { backgroundColor: theme?.colors?.white },
@@ -173,7 +174,7 @@ export class SearchBarIOS extends Component<
         ])}
       >
         <Input
-          testID="searchInput"
+          testID="RNE__SearchBar"
           renderErrorMessage={false}
           {...attributes}
           onFocus={this.onFocus}
@@ -240,27 +241,34 @@ export class SearchBarIOS extends Component<
           onLayout={(event) =>
             this.setState({ cancelButtonWidth: event.nativeEvent.layout.width })
           }
+          testID="RNE__SearchBar-cancelButtonContainer"
         >
-          <TouchableOpacity
+          <Pressable
             accessibilityRole="button"
             onPress={this.cancel}
             disabled={buttonDisabled}
             {...otherCancelButtonProps}
           >
-            <View style={[buttonStyle, buttonDisabled && buttonDisabledStyle]}>
+            <View
+              style={StyleSheet.flatten([
+                buttonStyle,
+                buttonDisabled && buttonDisabledStyle,
+              ])}
+              testID="RNE__SearchBar-cancelButton"
+            >
               <Text
-                style={[
+                style={StyleSheet.flatten([
                   styles.buttonTextStyle,
                   buttonColor && { color: buttonColor },
                   buttonTextStyle,
                   buttonDisabled &&
                     (buttonDisabledTextStyle || styles.buttonTextDisabled),
-                ]}
+                ])}
               >
                 {cancelButtonTitle}
               </Text>
             </View>
-          </TouchableOpacity>
+          </Pressable>
         </View>
       </View>
     );
