@@ -1,88 +1,68 @@
-/* eslint-disable jest/no-identical-title */
 import React from 'react';
-import { shallow } from 'enzyme';
-import { View } from 'react-native';
-import toJson from 'enzyme-to-json';
-import theme from '../../config/theme';
+import { fireEvent } from '@testing-library/react-native';
+import { ActivityIndicator, Text, Pressable, View } from 'react-native';
+import { renderWithWrapper } from '../../../.ci/testHelper';
+import { Icon } from '../../Icon';
+
+const WrapperTestID = 'RNE__SearchBar-wrapper';
+const SearchBarTestID = 'RNE__SearchBar';
 
 export function commonTests(SearchBar) {
-  it('should render without issues', () => {
-    const component = shallow(<SearchBar theme={theme} />);
-    expect(component.length).toBe(1);
-    expect(toJson(component)).toMatchSnapshot();
+  beforeEach(() => {
+    jest.resetModules();
   });
 
-  it('should render with a preset value', () => {
-    const component = shallow(<SearchBar theme={theme} value="Chickens" />);
-    expect(component.length).toBe(1);
-    expect(toJson(component)).toMatchSnapshot();
+  it('should match snapshot', () => {
+    const component = renderWithWrapper(<SearchBar />);
+    expect(component.toJSON()).toMatchSnapshot();
   });
 
   describe('Handlers', () => {
     it('onClear', () => {
-      const component = shallow(<SearchBar theme={theme} />);
-      component.find({ testID: 'searchInput' }).simulate('clear');
+      const handler = jest.fn();
+      const { queryByTestId } = renderWithWrapper(
+        <SearchBar onClear={handler} />
+      );
+      const Wrapper = queryByTestId(SearchBarTestID);
+      fireEvent(Wrapper, 'clear');
+      expect(handler).toBeCalled();
     });
 
     it('onFocus', () => {
-      const component = shallow(<SearchBar theme={theme} />);
-      component.find({ testID: 'searchInput' }).simulate('focus');
+      const handler = jest.fn();
+      const { queryByTestId } = renderWithWrapper(
+        <SearchBar onFocus={handler} />
+      );
+      const Wrapper = queryByTestId(SearchBarTestID);
+      fireEvent(Wrapper, 'focus');
+      expect(handler).toBeCalled();
     });
 
     it('onBlur', () => {
-      const component = shallow(<SearchBar theme={theme} />);
-      component.find({ testID: 'searchInput' }).simulate('blur');
+      const handler = jest.fn();
+      const { queryByTestId } = renderWithWrapper(
+        <SearchBar onBlur={handler} />
+      );
+      const Wrapper = queryByTestId(SearchBarTestID);
+      fireEvent(Wrapper, 'blur');
+      expect(handler).toBeCalled();
     });
 
-    it('onChangeText', () => {
-      const component = shallow(<SearchBar theme={theme} />);
-      component.find({ testID: 'searchInput' }).simulate('changeText', 'test');
-    });
-  });
-
-  describe('Instance methods', () => {
-    it('focus', () => {
-      const focus = jest.fn();
-      const component = shallow(<SearchBar theme={theme} />);
-      const instance = component.instance();
-      // Refs not available in shallow render
-      instance.input = {
-        focus,
-      };
-      instance.focus();
-      expect(focus).toHaveBeenCalledTimes(1);
-    });
-
-    it('blur', () => {
-      const blur = jest.fn();
-      const component = shallow(<SearchBar theme={theme} />);
-      const instance = component.instance();
-      // Refs not available in shallow render
-      instance.input = {
-        blur,
-      };
-      instance.blur();
-      expect(blur).toHaveBeenCalledTimes(1);
-    });
-
-    it('clear', () => {
-      const clear = jest.fn();
-      const component = shallow(<SearchBar theme={theme} />);
-      const instance = component.instance();
-      // Refs not available in shallow render
-      instance.input = {
-        clear,
-      };
-      instance.clear();
-      expect(clear).toHaveBeenCalledTimes(1);
+    it.skip('onChangeText', () => {
+      const handler = jest.fn();
+      const { queryByTestId } = renderWithWrapper(
+        <SearchBar onChangeText={handler} />
+      );
+      const Wrapper = queryByTestId(SearchBarTestID);
+      fireEvent.changeText(Wrapper, 'test');
+      expect(handler).toBeCalled();
     });
   });
 
   describe('Props', () => {
     it('showLoading, loadingProps', () => {
-      const component = shallow(
+      const { queryByTestId } = renderWithWrapper(
         <SearchBar
-          theme={theme}
           showLoading
           loadingProps={{
             style: { flex: 1 },
@@ -90,129 +70,126 @@ export function commonTests(SearchBar) {
           containerStyle={{ height: 70 }}
         />
       );
-      expect(component.length).toBe(1);
-      expect(toJson(component)).toMatchSnapshot();
+      const Wrapper = queryByTestId(WrapperTestID);
+      const component = Wrapper.findByType(ActivityIndicator);
+      expect(component.props.style.flex).toBe(1);
     });
 
     describe('searchIcon and without', () => {
       it('searchIcon', () => {
-        const component = shallow(
-          <SearchBar theme={theme} searchIcon={{ size: 50 }} lightTheme />
+        const { queryByTestId } = renderWithWrapper(
+          <SearchBar searchIcon={{ size: 50 }} lightTheme />
         );
-        expect(component.length).toBe(1);
-        expect(toJson(component)).toMatchSnapshot();
+        const Wrapper = queryByTestId(WrapperTestID);
+        const component = Wrapper.findByType(Icon);
+        expect(component.props.size).toBe(50);
       });
 
       it('custom searchIcon', () => {
-        const component = shallow(
-          <SearchBar theme={theme} searchIcon={<View />} round />
+        const { queryByTestId } = renderWithWrapper(
+          <SearchBar searchIcon={<View />} round />
         );
-        expect(component.length).toBe(1);
-        expect(toJson(component)).toMatchSnapshot();
+        const Wrapper = queryByTestId(WrapperTestID);
+        expect(Wrapper.findAllByType(View)).not.toBeNull();
       });
 
       it('no searchIcon', () => {
-        const component = shallow(
-          <SearchBar theme={theme} searchIcon={false} />
+        const { queryByTestId } = renderWithWrapper(
+          <SearchBar searchIcon={false} />
         );
-        expect(component.length).toBe(1);
-        expect(toJson(component)).toMatchSnapshot();
+        const Wrapper = queryByTestId(SearchBarTestID);
+        expect(Wrapper.props.leftIcon).toBeFalsy();
       });
     });
 
     describe('clearIcon and without', () => {
-      it('clearIcon', () => {
-        const component = shallow(
-          <SearchBar theme={theme} clearIcon={{ color: 'black' }} />
+      it.skip('clearIcon', () => {
+        const { queryByTestId } = renderWithWrapper(
+          <SearchBar
+            searchIcon={false}
+            value="test"
+            clearIcon={{ color: 'black' }}
+          />
         );
-        expect(component.length).toBe(1);
-        expect(toJson(component)).toMatchSnapshot();
+        const Wrapper = queryByTestId(WrapperTestID);
+        const component = Wrapper.findByType(Icon);
+        expect(component.props.color).toBe('black');
       });
 
       it('custom clearIcon', () => {
-        const component = shallow(
-          <SearchBar theme={theme} clearIcon={<View />} />
+        const { queryByTestId } = renderWithWrapper(
+          <SearchBar clearIcon={<View />} />
         );
-        expect(component.length).toBe(1);
-        expect(toJson(component)).toMatchSnapshot();
+        const Wrapper = queryByTestId(WrapperTestID);
+        expect(Wrapper.findAllByType(View)).not.toBeNull();
       });
 
       it('no clearIcon', () => {
-        const component = shallow(
-          <SearchBar theme={theme} clearIcon={false} />
+        const { queryByTestId } = renderWithWrapper(
+          <SearchBar clearIcon={false} />
         );
-        expect(component.length).toBe(1);
-        expect(toJson(component)).toMatchSnapshot();
+        const Wrapper = queryByTestId(WrapperTestID);
+        expect(Wrapper.props.leftIcon).toBeFalsy();
       });
     });
   });
 }
 
 export function commonPlatformTest(SearchBar) {
-  describe('Handlers', () => {
+  describe('Platform Handlers', () => {
     it('onCancel', () => {
-      const component = shallow(<SearchBar theme={theme} />);
-      component.find({ testID: 'searchInput' }).simulate('cancel');
-    });
-  });
-
-  describe('Instance methods', () => {
-    it('cancel', () => {
-      const onCancel = jest.fn();
-      const component = shallow(
-        <SearchBar theme={theme} onCancel={onCancel} />
+      const handler = jest.fn();
+      const { queryByTestId } = renderWithWrapper(
+        <SearchBar onCancel={handler} />
       );
-      const instance = component.instance();
-      // Refs not available in shallow render
-      instance.input = {
-        blur: jest.fn(),
-      };
-      instance.cancel();
-      jest.runAllTimers();
-      expect(onCancel).toHaveBeenCalledTimes(1);
+      const Wrapper = queryByTestId(SearchBarTestID);
+      fireEvent(Wrapper, 'cancel');
+      expect(handler).toBeCalled();
     });
   });
 
-  describe('Props', () => {
+  describe('Platform Props', () => {
     describe('cancel button', () => {
       describe('Enabled', () => {
         it('cancelButtonTitle', () => {
-          const component = shallow(
-            <SearchBar theme={theme} cancelButtonTitle="Annuler" />
+          const { queryByTestId } = renderWithWrapper(
+            <SearchBar cancelButtonTitle="Annuler" />
           );
-          expect(component.length).toBe(1);
-          expect(toJson(component)).toMatchSnapshot();
+          const Wrapper = queryByTestId('RNE__SearchBar-cancelButton');
+          expect(Wrapper?.findByType(Text).props.children).toBe('Annuler');
         });
 
         it('cancelButtonProps', () => {
-          const component = shallow(
-            <SearchBar
-              theme={theme}
-              cancelButtonProps={{
-                color: 'black',
-                buttonStyle: { elevation: 0 },
-                buttonTextStyle: { fontSize: 12 },
-              }}
-            />
+          const Props = {
+            color: 'black',
+            buttonStyle: { elevation: 0 },
+            buttonTextStyle: { fontSize: 12 },
+          };
+          const { queryByTestId } = renderWithWrapper(
+            <SearchBar cancelButtonProps={Props} />
           );
-          expect(component.length).toBe(1);
-          expect(toJson(component)).toMatchSnapshot();
+          const Wrapper = queryByTestId('RNE__SearchBar-cancelButton');
+          expect(Wrapper.props.style).toMatchObject({
+            elevation: 0,
+          });
+          expect(Wrapper.findByType(Text).props.style).toMatchObject({
+            color: 'black',
+            fontSize: 12,
+          });
         });
       });
 
       describe('Disabled', () => {
         it('cancelButtonProps', () => {
-          const component = shallow(
-            <SearchBar cancelButtonProps={{ disabled: true }} theme={theme} />
+          const { queryByTestId } = renderWithWrapper(
+            <SearchBar cancelButtonProps={{ disabled: true }} />
           );
-          expect(component.length).toBe(1);
-          expect(toJson(component)).toMatchSnapshot();
+          const Wrapper = queryByTestId('RNE__SearchBar-cancelButtonContainer');
+          expect(Wrapper.findByType(Pressable).props.disabled).toBeTruthy();
         });
-
         it('cancelButtonProps disabled styles', () => {
-          const component = shallow(
+          const { queryByTestId } = renderWithWrapper(
             <SearchBar
-              theme={theme}
               cancelButtonProps={{
                 disabled: true,
                 buttonDisabledStyle: { backgroundColor: '#cdcdcd' },
@@ -220,8 +197,13 @@ export function commonPlatformTest(SearchBar) {
               }}
             />
           );
-          expect(component.length).toBe(1);
-          expect(toJson(component)).toMatchSnapshot();
+          const Wrapper = queryByTestId('RNE__SearchBar-cancelButton');
+          expect(Wrapper.props.style).toMatchObject({
+            backgroundColor: '#cdcdcd',
+          });
+          expect(Wrapper.findByType(Text).props.style).toMatchObject({
+            color: '#ffffff',
+          });
         });
       });
     });
