@@ -48,33 +48,79 @@ type Sizable = {
 };
 
 export type SliderProps = {
+  /** Initial value of the slider. */
   value?: number;
+
+  /** If true the user won't be able to move the slider. */
   disabled?: boolean;
+
+  /** Initial minimum value of the slider. */
   minimumValue?: number;
+
+  /** Initial maximum value of the slider. */
   maximumValue?: number;
+
+  /** Step value of the slider. The value should be between 0 and maximumValue - minimumValue). */
   step?: number;
+
+  /** The color used for the track to the left of the button. */
   minimumTrackTintColor?: string;
+
+  /** The color used for the track to the right of the button. */
   maximumTrackTintColor?: string;
+
+  /** If true, thumb will respond and jump to any touch along the track. */
   allowTouchTrack?: boolean;
+
+  /** The color used for the thumb. */
   thumbTintColor?: string;
+
+  /** The size of the touch area that allows moving the thumb. The touch area has the same center as the visible thumb. This allows to have a visually small thumb while still allowing the user to move it easily. */
   thumbTouchSize?: Sizable;
+
+  /** Callback continuously called while the user is dragging the slider. */
   onValueChange?(value: number): void;
+
+  /** Callback called when the user starts changing the value (e.g. when the slider is pressed). */
   onSlidingStart?(value: number): void;
+
+  /** Callback called when the user finishes changing the value (e.g. when the slider is released). */
   onSlidingComplete?(value: number): void;
+
+  /** The style applied to the slider container. */
   style?: StyleProp<ViewStyle>;
+
+  /** The style applied to the track. */
   trackStyle?: StyleProp<ViewStyle>;
+
+  /** The style applied to the thumb. */
   thumbStyle?: StyleProp<ViewStyle>;
+
+  /** The props applied to the thumb. Uses `Component` prop which can accept `Animated` components. */
   thumbProps?: any;
+
+  /** Set this to true to visually see the thumb touch rect in green. */
   debugTouchArea?: boolean;
+
+  /** Set to true if you want to use the default 'spring' animation. */
   animateTransitions?: boolean;
+
+  /** Set to 'spring' or 'timing' to use one of those two types of animations with the default [animation properties](https://reactnative.dev/docs/animations.html). */
   animationType?: 'spring' | 'timing';
+
+  /** Set the orientation of the slider. */
   orientation?: 'horizontal' | 'vertical';
+
+  /** Used to configure the animation parameters. These are the same parameters in the [Animated library](https://reactnative.dev/docs/animations.html). */
   animationConfig?:
     | Animated.TimingAnimationConfig
     | Animated.SpringAnimationConfig;
+
+  /** Apply style to the container of the slider. */
   containerStyle?: typeof styles;
 };
 
+/** Sliders allow users to select a value from a fixed set of values using drag utility.*/
 export const Slider: RneFunctionComponent<SliderProps> = ({
   minimumValue,
   maximumValue,
@@ -454,6 +500,7 @@ export const Slider: RneFunctionComponent<SliderProps> = ({
 
   return (
     <View
+      testID="RNE__Slider_Container"
       {...other}
       style={StyleSheet.flatten([
         isVertical.current
@@ -466,10 +513,11 @@ export const Slider: RneFunctionComponent<SliderProps> = ({
       accessibilityValue={{
         min: minimumValue,
         max: maximumValue,
-        now: props.value,
+        now: currentPropValue.current,
       }}
     >
       <View
+        testID="RNE__Slider_Track_maximum"
         style={StyleSheet.flatten([
           mainStyles.track,
           isVertical.current
@@ -482,6 +530,7 @@ export const Slider: RneFunctionComponent<SliderProps> = ({
       />
 
       <Animated.View
+        testID="RNE__Slider_Track_minimum"
         style={StyleSheet.flatten([
           mainStyles.track,
           isVertical.current
@@ -501,6 +550,7 @@ export const Slider: RneFunctionComponent<SliderProps> = ({
         {...thumbProps}
       />
       <View
+        testID="RNE__Slider_TouchArea"
         style={StyleSheet.flatten([styles.touchArea, touchOverflowStyle])}
         {...panResponder.panHandlers}
       >
