@@ -8,16 +8,26 @@ import {
   StyleSheet,
 } from 'react-native';
 import { RneFunctionComponent } from '../helpers';
-import { TabItemProps } from './TabItem';
+import { TabItemProps } from './Tab.Item';
 
 export type TabBaseProps = ViewProps & {
+  /** Child position index value. */
   value?: number;
+
+  /** On Index Change Callback. */
   onChange?: (value: number) => void;
+
+  /** Disable the indicator below. */
   disableIndicator?: boolean;
+
+  /** Additional styling for tab indicator. */
   indicatorStyle?: StyleProp<ViewStyle>;
+
+  /** Define the background Variant. */
   variant?: 'primary' | 'default';
 };
 
+/** Tabs organize content across different screens, data sets, and other interactions. */
 export const TabBase: RneFunctionComponent<TabBaseProps> = ({
   theme,
   children,
@@ -26,7 +36,7 @@ export const TabBase: RneFunctionComponent<TabBaseProps> = ({
   indicatorStyle,
   disableIndicator,
   variant,
-  ...props
+  ...rest
 }) => {
   const [dim, setDim] = React.useState({ width: 0 });
   const { current: animation } = React.useRef(new Animated.Value(0));
@@ -44,7 +54,7 @@ export const TabBase: RneFunctionComponent<TabBaseProps> = ({
   return (
     <>
       <View
-        {...props}
+        {...rest}
         accessibilityRole="tablist"
         style={[
           styles.viewStyle,
@@ -113,4 +123,4 @@ const styles = StyleSheet.create({
   },
 });
 
-TabBase.displayName = 'TabBase';
+TabBase.displayName = 'Tab';
