@@ -2,7 +2,7 @@ import { withTheme } from '../config';
 import { RneFunctionComponent } from '../helpers';
 
 import { TabViewBase, TabViewBaseProps } from './TabView';
-import { TabViewItem } from './TabViewItem';
+import { TabViewItem, TabViewItemProps } from './TabView.Item';
 
 const ThemedTabViewItem = withTheme(TabViewItem, 'TabViewItem');
 
@@ -12,7 +12,7 @@ export type TabViewProps = RneFunctionComponent<TabViewBaseProps> & {
 
 export const TabView: TabViewProps = Object.assign(TabViewBase);
 
-export type { TabViewBaseProps };
+export type { TabViewBaseProps, TabViewItemProps };
 
 export const ThemedTabView = Object.assign(withTheme(TabView, 'TabView'), {
   Item: ThemedTabViewItem,
