@@ -3,9 +3,8 @@ import {
   View,
   StyleSheet,
   Dimensions,
-  TouchableOpacity,
-  TouchableOpacityProps,
-  TouchableNativeFeedbackProps,
+  Pressable,
+  PressableProps,
   ViewStyle,
   StyleProp,
   TextStyle,
@@ -17,30 +16,67 @@ import Image from '../Image';
 import Text from '../Text';
 import Icon, { IconObject, IconProps } from '../Icon';
 import { ThemedFeaturedTile } from './components/FeaturedTile';
-import { RneFunctionComponent } from '../helpers';
+import { androidRipple, RneFunctionComponent } from '../helpers';
+import Color from 'color';
 
-export type TileProps = TouchableOpacityProps &
-  TouchableNativeFeedbackProps & {
-    title?: string;
-    icon?: IconObject & IconProps;
-    caption?: React.ReactNode;
-    imageSrc?: ImageSourcePropType | string | number;
-    activeOpacity?: number;
-    containerStyle?: StyleProp<ViewStyle>;
-    imageContainerStyle?: StyleProp<ViewStyle>;
-    iconContainerStyle?: StyleProp<ViewStyle>;
-    overlayContainerStyle?: StyleProp<ViewStyle>;
-    titleStyle?: StyleProp<TextStyle>;
-    captionStyle?: StyleProp<TextStyle>;
-    width?: number;
-    height?: number;
-    featured?: boolean;
-    contentContainerStyle?: StyleProp<ViewStyle>;
-    titleNumberOfLines?: number;
-    imageProps?: Partial<ImageProps>;
-    ImageComponent?: typeof React.Component;
-  };
+export type TileProps = PressableProps & {
+  /** Text inside the tile. */
+  title?: string;
 
+  /** Icon Component Props. */
+  icon?: IconObject & IconProps;
+
+  /** Text inside the tilt when tile is featured. */
+  caption?: React.ReactNode;
+
+  /** Source for the image. */
+  imageSrc?: ImageSourcePropType | string | number;
+
+  /** Number passed to control opacity on press. */
+  activeOpacity?: number;
+
+  /** Styling for the outer tile container. */
+  containerStyle?: StyleProp<ViewStyle>;
+
+  /** Styling for the image. */
+  imageContainerStyle?: StyleProp<ViewStyle>;
+
+  /** Styling for the outer icon container. */
+  iconContainerStyle?: StyleProp<ViewStyle>;
+
+  /** Styling for the overlay container when using featured tile. */
+  overlayContainerStyle?: StyleProp<ViewStyle>;
+
+  /** Styling for the title. */
+  titleStyle?: StyleProp<TextStyle>;
+
+  /** Styling for the caption (optional); You only use this if `caption` is a string. */
+  captionStyle?: StyleProp<TextStyle>;
+
+  /** Width for the tile. */
+  width?: number;
+
+  /** Height for the tile. */
+  height?: number;
+
+  /** Changes the look of the tile. */
+  featured?: boolean;
+
+  /** Styling for bottom container when not featured tile. */
+  contentContainerStyle?: StyleProp<ViewStyle>;
+
+  /** Number of lines for Title. */
+  titleNumberOfLines?: number;
+
+  /** Optional properties to pass to the image if provided e.g "resizeMode". */
+  imageProps?: Partial<ImageProps>;
+
+  /** Custom ImageComponent for Tile. */
+  ImageComponent?: typeof React.Component;
+};
+
+/** Tiles like Cards, are a convenient way to display related content about a single subject.
+ * Also receives all [TouchableNativeFeedback](http://reactnative.dev/docs/touchablenativefeedback.html#props) (Android) or [TouchableOpacity](http://reactnative.dev/docs/touchableopacity.html#props) (iOS) props. */
 export const Tile: RneFunctionComponent<TileProps> = ({
   featured,
   imageSrc,
@@ -62,6 +98,7 @@ export const Tile: RneFunctionComponent<TileProps> = ({
   imageProps = {},
   width = Dimensions.get('window').width,
   height = width * 0.8,
+  theme,
   ...attributes
 }) => {
   if (featured) {
@@ -86,10 +123,14 @@ export const Tile: RneFunctionComponent<TileProps> = ({
   }
 
   return (
-    <TouchableOpacity
-      {...attributes}
-      onPress={onPress}
-      activeOpacity={activeOpacity}
+    <Pressable
+      {...{
+        onPress,
+        android_ripple: androidRipple(
+          Color(theme?.colors?.primary).alpha(activeOpacity).rgb().toString()
+        ),
+        ...attributes,
+      }}
       style={StyleSheet.flatten([
         {
           width,
@@ -138,7 +179,7 @@ export const Tile: RneFunctionComponent<TileProps> = ({
         </Text>
         {children}
       </View>
-    </TouchableOpacity>
+    </Pressable>
   );
 };
 
