@@ -7,14 +7,13 @@ export const useTheme = () => {
   return useContext(ThemeContext);
 };
 
-export const makeStyles = <
-  T extends StyleSheet.NamedStyles<T> | StyleSheet.NamedStyles<any>,
-  V
->(
-  styles: T | ((theme: Partial<FullTheme>, props: V) => T)
-) => (props: V = {} as any): T => {
-  const { theme } = useTheme();
-  const css = typeof styles === 'function' ? styles(theme, props) : styles;
+export const makeStyles =
+  <T extends StyleSheet.NamedStyles<T> | StyleSheet.NamedStyles<any>, V>(
+    styles: T | ((theme: Partial<FullTheme>, props: V) => T)
+  ) =>
+  (props: V = {} as any): T => {
+    const { theme } = useTheme();
+    const css = typeof styles === 'function' ? styles(theme, props) : styles;
 
-  return StyleSheet.create(css);
-};
+    return StyleSheet.create(css);
+  };
