@@ -1,4 +1,4 @@
-import { Platform, Dimensions } from 'react-native';
+import { Platform, Dimensions, PressableProps, ColorValue } from 'react-native';
 import color from 'color';
 import renderNode from './renderNode';
 import getIconType from './getIconType';
@@ -13,6 +13,16 @@ const isIOS = Platform.OS === 'ios';
 export type RneFunctionComponent<T> = React.FunctionComponent<
   T & Partial<ThemeProps<T>>
 >;
+
+export const androidRipple = (
+  rippleColor?: string | ColorValue
+): PressableProps['android_ripple'] => {
+  return {
+    borderless: false,
+    color: rippleColor,
+    radius: -5,
+  };
+};
 
 export const patchWebProps = <T extends Record<any, any>>({
   updateTheme,
@@ -32,3 +42,14 @@ export {
   isIOS,
   color,
 };
+
+type Inline<T, K extends keyof T> = Partial<
+  {
+    pressableProps: Omit<T, K>;
+  } & Pick<T, K>
+>;
+
+export type InlinePressableProps = Inline<
+  PressableProps,
+  'onPress' | 'onLongPress' | 'onPressIn' | 'onPressOut'
+>;
